pyproject.toml
[project]
name = "winclaw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.91.0",
    "chardet>=7.0.1",
    "croniter>=6.0.0",
    "dingtalk-stream>=0.24.3",
    "fastmcp>=3.1.0",
    "httpx>=0.28.1",
    "json-repair>=0.58.5",
    "lark-oapi>=1.5.3",
    "litellm>=1.82.0",
    "loguru>=0.7.3",
    "markdownify>=1.2.2",
    "mcp>=1.26.0",
    "msgpack>=1.1.2",
    "openai>=2.26.0",
    "pillow>=12.1.1",
    "playwright>=1.58.0",
    "prompt-toolkit>=3.0.52",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyperclip>=1.11.0",
    "pyyaml>=6.0.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "python-socks[asyncio]>=2.8.1",
    "pywinauto>=0.6.9",
    "qq-botpy>=1.2.1",
    "ruff>=0.15.5",
    "sqlite-vec>=0.1.6",
    "tavily>=1.1.0",
    "tiktoken>=0.12.0",
    "typer>=0.24.1",
    "websockets>=16.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# Example agent tools use PascalCase filenames (e.g. FileReadTool.py) per project convention.
"examples/tools/*.py" = ["N999"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project.scripts]
winclaw = "winclaw.cli.commands:app"
